rename renderMovie to loadMoviePage.js, delete block of code that returned new event, import loadMovie fun to movie.js
